Fixed implementation

+++ payment/app.py
@@ -40,24 +40,46 @@ def add_credit(user_id: str, amount: int):
         return {'Error': 'Amount must be positive'}, 400
     user = user_collection.find_one_or_404({'_id': ObjectId(user_id)})
     credit: int = int(user['credit'])
-    user['credit'] = credit + int(amount)
-    user_collection.update_one(user)
+    newCredit = credit + int(amount)
+    result = user_collection.update_one(
+        # Filter to check if the credit is unchanged
+        {'_id': ObjectId(user_id), 'credit': { '$eq': credit}},
+        {'$set': {'credit': newCredit}}
+    )
+
+    if result.modified_count != 1:
+        return {'Error': 'The credit was updated by someone else'}, 400
+    
+    return {'Success': 'Credit is updated successfully'}, 200
+
 
 
 @app.post('/pay/<user_id>/<order_id>/<amount>')
 def remove_credit(user_id: str, order_id: str, amount: int):
-    user = user_collection.find_one_or_404({'_id': ObjectId(user_id)})
-    if user['credit'] < int(amount):
-        return 'Not enough credit', 400
     
     order = paid_order_colleciton.find_one({'order_id': order_id})
 
     if order:
         return 'Order already paid', 401
 
+    user = user_collection.find_one_or_404({'_id': ObjectId(user_id)})
+    credit: int = int(user['credit'])
+    
+    # For some reason amount is not an int but string at this point
+    if credit < int(amount):
+        return 'Not enough credit', 400
+
     # Update to transaction
-    user['credit'] -= amount
-    user_collection.update_one(user)
+    newCredit = credit - int(amount)
+    result = user_collection.update_one(
+        # Filter to check if the credit is unchanged
+        {'_id': ObjectId(user_id), 'credit': { '$eq': credit}},
+        {'$set': {'credit': newCredit}}
+    )
+
+    if result.modified_count != 1:
+        return {'Error': 'Something went wrong checking out the order'}, 400
+
     paid_order_colleciton.insert_one({'order_id': order_id, 'amount': amount}) 
     return 'Success', 200
 
@@ -65,12 +87,21 @@ def remove_credit(user_id: str, order_id: str, amount: int):
 @app.post('/cancel/<user_id>/<order_id>')
 def cancel_payment(user_id: str, order_id: str):
     order = paid_order_colleciton.find_one_or_404({'order_id': order_id})
+    amount = int(order['amount'])
+    
     user = user_collection.find_one_or_404({'_id': ObjectId(user_id)})
-    amount = order['amount']
+    credit = int(user['credit'])
 
-    user['credit'] += amount
+    newCredit = credit + amount
+    result = user_collection.update_one(
+        # Filter to check if the credit is unchanged
+        {'_id': ObjectId(user_id), 'credit': { '$eq': credit}},
+        {'$set': {'credit': newCredit}}
+    )
 
-    user_collection.update_one(user)
+    if result.modified_count != 1:
+        return {'Error': 'Something went wrong checking cancelling the order'}, 400
+
     paid_order_colleciton.delete_one({'order_id': order_id})
     return 'Success', 200
